Revert "Revert "Move some folders to protected""
This reverts commit 725ba37afd8eb2d06169590adfb8499a245cfd76.
