+++ frontend/app/[locale]/(protected)/payment/[transaction_id]/page.tsx
@@ -1,6 +1,7 @@
 import { Metadata } from "next"
-import PaymentDetailsClient from "./payment-details-client"
+import PaymentClient from "./payment-client"
 import { getTranslations } from "next-intl/server"
+import { getStripedataWalletStripedataGet } from "../../../../../src/codegen"
 
 interface Props {
   params: Promise<{
@@ -14,7 +15,7 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
   const t = await getTranslations()
 
   return {
-    title: t("transaction-summary"),
+    title: t("payment"),
     robots: {
       index: false,
       follow: false,
@@ -22,7 +23,14 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
   }
 }
 
-export default async function PaymentDetailsPage({ params }: Props) {
+export default async function PaymentPage({ params }: Props) {
   const { transaction_id } = await params
-  return <PaymentDetailsClient transactionId={transaction_id} />
+  const stripeDataQuery = await getStripedataWalletStripedataGet()
+
+  return (
+    <PaymentClient
+      transactionId={transaction_id}
+      stripePublicKey={stripeDataQuery.data.public_key}
+    />
+  )
 }
