+++ frontend/app/[locale]/(protected)/payment/details/[transaction_id]/page.tsx
@@ -1,0 +1,27 @@
+import { Metadata } from "next"
+import PaymentDetailsClient from "./payment-details-client"
+import { getTranslations } from "next-intl/server"
+
+interface Props {
+  params: Promise<{
+    transaction_id: string
+    locale: string
+  }>
+}
+
+export async function generateMetadata(): Promise<Metadata> {
+  const t = await getTranslations()
+
+  return {
+    title: t("transaction-summary"),
+    robots: {
+      index: false,
+      follow: false,
+    },
+  }
+}
+
+export default async function PaymentDetailsPage({ params }: Props) {
+  const { transaction_id } = await params
+  return <PaymentDetailsClient transactionId={transaction_id} />
+}
