+++ frontend/app/[locale]/(protected)/purchase/checkout/checkout-client.tsx
