+++ frontend/app/[locale]/(protected)/purchase/checkout/page.tsx
