+++ frontend/app/[locale]/(protected)/purchase/finished/page.tsx
