+++ frontend/app/[locale]/(protected)/purchase/finished/purchase-finished-client.tsx
