+++ frontend/app/[locale]/(protected)/purchase/page.tsx
